+++ test/datasets/test_numpy.py
@@ -11,7 +11,8 @@ from torcheeg.datasets import NumpyDataset
 class TestNumpyDataset(unittest.TestCase):
 
     def setUp(self):
-        shutil.rmtree('./tmp_out/')
+        if os.path.exists('./tmp_out/'):
+            shutil.rmtree('./tmp_out/')
         os.mkdir('./tmp_out/')
 
     def test_numpy_dataset(self):
